move things to drop_condition
