+++ test-loop-tests/src/setup/env.rs
@@ -1,15 +1,7 @@
-use near_async::messaging::{CanSend, LateBoundSender};
 use near_async::test_loop::TestLoopV2;
 use near_async::test_loop::data::TestLoopData;
-use near_async::test_loop::sender::TestLoopSender;
 use near_async::time::Duration;
-use near_chunks::adapter::ShardsManagerRequestFromClient;
-use near_chunks::shards_manager_actor::ShardsManagerActor;
-use near_primitives::sharding::{ChunkHash, ShardChunkHeader};
-use near_primitives_core::types::BlockHeight;
-use std::collections::HashMap;
 use std::sync::atomic::Ordering;
-use std::sync::{Arc, Mutex};
 use tempfile::TempDir;
 
 use super::drop_condition::DropCondition;
@@ -30,7 +22,7 @@ impl TestLoopEnv {
     /// While adding a new node to the environment, we can iterate through all the drop_conditions
     /// and register them with the new node's peer_manager_actor.
     pub fn drop(mut self, drop_condition: DropCondition) -> Self {
-        for data in self.node_datas.iter_mut() {
+        for data in self.node_datas.iter() {
             data.register_drop_condition(
                 &mut self.test_loop.data,
                 self.shared_state.chunks_storage.clone(),
@@ -95,63 +87,5 @@ impl TestLoopEnv {
 
         self.test_loop.shutdown_and_drain_remaining_events(timeout);
         self.shared_state.tempdir
-    }
-}
-
-/// Stores all chunks ever observed on chain. Determines if a chunk can be
-/// dropped within a test loop.
-///
-/// Needed to intercept network messages storing chunk hash only, while
-/// interception requires more detailed information like shard id.
-#[derive(Default)]
-pub struct TestLoopChunksStorage {
-    /// Mapping from chunk hashes to headers.
-    storage: HashMap<ChunkHash, ShardChunkHeader>,
-    /// Minimal chunk height ever observed.
-    min_chunk_height: Option<BlockHeight>,
-}
-
-impl TestLoopChunksStorage {
-    pub fn insert(&mut self, chunk_header: ShardChunkHeader) {
-        let chunk_height = chunk_header.height_created();
-        self.min_chunk_height = Some(
-            self.min_chunk_height
-                .map_or(chunk_height, |current_height| current_height.min(chunk_height)),
-        );
-        self.storage.insert(chunk_header.chunk_hash(), chunk_header);
-    }
-
-    pub fn get(&self, chunk_hash: &ChunkHash) -> Option<&ShardChunkHeader> {
-        self.storage.get(chunk_hash)
-    }
-
-    /// If chunk height is too low, don't drop chunk, allow the chain to warm
-    /// up.
-    pub fn can_drop_chunk(&self, chunk_header: &ShardChunkHeader) -> bool {
-        self.min_chunk_height
-            .is_some_and(|min_height| chunk_header.height_created() >= min_height + 3)
-    }
-}
-
-/// Custom implementation of `Sender` for messages from `Client` to
-/// `ShardsManagerActor` that allows to intercept all messages indicating
-/// any chunk production and storing all chunks.
-pub struct ClientToShardsManagerSender {
-    pub sender: Arc<LateBoundSender<TestLoopSender<ShardsManagerActor>>>,
-    /// Storage of chunks shared between all test loop nodes.
-    pub chunks_storage: Arc<Mutex<TestLoopChunksStorage>>,
-}
-
-impl CanSend<ShardsManagerRequestFromClient> for ClientToShardsManagerSender {
-    fn send(&self, message: ShardsManagerRequestFromClient) {
-        // `DistributeEncodedChunk` indicates that a certain chunk was produced.
-        if let ShardsManagerRequestFromClient::DistributeEncodedChunk { partial_chunk, .. } =
-            &message
-        {
-            let mut chunks_storage = self.chunks_storage.lock().unwrap();
-            chunks_storage.insert(partial_chunk.cloned_header());
-        }
-        // After maybe storing the chunk, send the message as usual.
-        self.sender.send(message);
     }
 }
