+++ test-loop-tests/src/utils/network.rs
@@ -1,10 +1,11 @@
-use crate::setup::env::TestLoopChunksStorage;
 use near_epoch_manager::EpochManagerAdapter;
 use near_network::types::NetworkRequests;
 use near_primitives::sharding::ShardChunkHeader;
 use near_primitives::types::{AccountId, BlockHeight};
 use std::collections::HashSet;
 use std::sync::{Arc, Mutex};
+
+use crate::setup::drop_condition::TestLoopChunksStorage;
 
 type DropChunkCondition = Box<dyn Fn(ShardChunkHeader) -> bool>;
